Added state file system and new states for game story
