--- Assets/Room 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fb60c9f30fafa5848b2faa09365ac12c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
