--- Assets/Room 2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aadb39c377cb761428625d213be412c0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
