+++ Assets/Scripts/AdventureGame.cs
@@ -1,0 +1,48 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class AdventureGame : MonoBehaviour
+{
+
+    [SerializeField] Text textComponent; // SerializeField makes textcomponent available in Unity engine
+    // Start is called before the first frame update
+    [SerializeField] State startingState;
+
+    State state;
+
+    void Start()
+    {
+        state = startingState;
+        textComponent.text = state.GetStateStory();
+
+         
+    }
+     
+    // Update is called once per frame
+    void Update()
+    {
+        ManageState();
+    }
+
+    private void ManageState()
+    {
+        var nextStates = state.GetNextStates();
+
+        if (Input.GetKeyDown(KeyCode.Alpha1))
+        {
+            state = nextStates[0];
+        }
+        else if (Input.GetKeyDown(KeyCode.Alpha2))
+        {
+            state = nextStates[1];
+        }
+        else if (Input.GetKeyDown(KeyCode.Alpha3))
+        {
+            state = nextStates[2];
+            // Debug.Log("Invalid input");
+        }
+        textComponent.text = state.GetStateStory();
+    }
+}
