+++ Assets/Scripts/AdventureGame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b8e461a3901aad49bb0a5d22142ebc5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
