--- Assets/StartingState.asset
@@ -1,25 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8902a5de67c4b0944b75badcd67a421d, type: 3}
-  m_Name: StartingState
-  m_EditorClassIdentifier: 
-  storyText: "The protestors finished their march at the entrance to the BigCorp
-    headquarters.\r\n\rThey were mostly ignored by passersby but the plain clothes
-    agents circled around them, their formation every closing about the ignorant
-    prey.\r\n\rI watched for a time as one agent, standing at a distance from the
-    rest, signalled to them. Doubtless he was the Captain of the team, although dressed
-    as a civilian his gilded implant gave away his seniority.\r An explosion sounded
-    somewhere in the atmosphere, but it went mostly ignored by the population.\r\n\rWhat
-    do you do?\r\n1. Watch the security operation playout\r\n2. Start your mission"
-  nextStates:
-  - {fileID: 11400000, guid: fb60c9f30fafa5848b2faa09365ac12c, type: 2}
-  - {fileID: 11400000, guid: aadb39c377cb761428625d213be412c0, type: 2}
