--- Assets/StartingState.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 10dc1835b009379489d2cee7a9f36d22
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
