--- Assets/State.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8902a5de67c4b0944b75badcd67a421d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
