+++ Assets/States/Game Over/X1. Game Over.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 46b6a86d15953934fa455e39cc4cc235
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
