+++ Assets/States/Intro/A0. Introduction.asset
@@ -1,0 +1,33 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8902a5de67c4b0944b75badcd67a421d, type: 3}
+  m_Name: A0. Introduction
+  m_EditorClassIdentifier: 
+  storyText: 'Welcome to this game.
+
+
+    This is a chose your own adventure. Read
+    carefully and enjopy the experience!
+
+
+    At each stage you will be asked to
+    enter a number to do an action. At any point in time you can press Q to quit.
+
+
+    Press
+    1 to begin.
+
+
+
+'
+  nextStates:
+  - {fileID: 11400000, guid: 10dc1835b009379489d2cee7a9f36d22, type: 2}
