+++ Assets/States/Start + Inside Building.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5aeaa45dba87d6d4fa1300a142572695
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
