+++ Assets/States/Start + Inside Building/A1. Start.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10dc1835b009379489d2cee7a9f36d22
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
