+++ Assets/States/Start + Inside Building/A2. Security Operation.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 803566b39c194e84db64bb5c29cfeb57
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
