+++ Assets/States/Start + Inside Building/B1. Slip into HQ.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f4a67f8b4048f141b90f7084be1ad6a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
