+++ Assets/States/Start + Inside Building/B2. Guard Barracks.asset
@@ -1,0 +1,28 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8902a5de67c4b0944b75badcd67a421d, type: 3}
+  m_Name: B2. Guard Barracks
+  m_EditorClassIdentifier: 
+  storyText: 'You dash across the foyer to the guard barracks, slipping through the
+    open door.
+
+    
+
+    Inside its vacant of guards. Thinking quickly you raid
+    an opened guards locker, changing into their uniform and grabbing their keys.
+
+    
+
+    1.
+    Continue...'
+  nextStates:
+  - {fileID: 11400000, guid: 693a7b690d7a5434ebb2502d98adb2ee, type: 2}
