+++ Assets/States/Start + Inside Building/B2. Guard Barracks.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c7253c0cc28dba8448b460c6207f86e2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
