+++ Assets/States/Start + Inside Building/B3. Fire stairs.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d3eaa6aa21a9754095f2634f1f4cb7f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
