+++ Assets/States/Start + Inside Building/B4. Foyer.asset
@@ -1,0 +1,32 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8902a5de67c4b0944b75badcd67a421d, type: 3}
+  m_Name: B4. Foyer
+  m_EditorClassIdentifier: 
+  storyText: 'Back in the still empty foyer two options lay ahead: ascend with the
+    fire stairs or use the security card to take the elevator.
+
+    
+
+    How do
+    you proceed?
+
+    
+
+    1. Take the fire stairs
+
+    
+
+    2. Use the elevator'
+  nextStates:
+  - {fileID: 11400000, guid: 4d3eaa6aa21a9754095f2634f1f4cb7f, type: 2}
+  - {fileID: 11400000, guid: 71b1e59b7714af5488b7d798c06d794c, type: 2}
