+++ Assets/States/Start + Inside Building/B4. Foyer.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 693a7b690d7a5434ebb2502d98adb2ee
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
