+++ Assets/States/Start + Inside Building/C1. Elevator.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71b1e59b7714af5488b7d798c06d794c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
