+++ Assets/States/Start + Inside Building/C2. Guess Passcode.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c9b27b73254e72458a931efd45a933b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
