+++ Assets/States/Start + Inside Building/C3. Choose a Different Floor.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8902a5de67c4b0944b75badcd67a421d, type: 3}
+  m_Name: C3. Choose a Different Floor
+  m_EditorClassIdentifier: 
+  storyText: 'You try a number of different floors till you finally get one to work:
+    floor 26. Close enough to the 32nd floor. Recalling your briefing you remember
+    that floor 26 is a workers quarters for overnight maintenance and cleaning staff
+    that should be mostly empty at this time. '
+  nextStates: []
