+++ Assets/States/Start + Inside Building/C3. Choose a Different Floor.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff78d5a658af4d049825928ae6b6b400
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
